Alterei o tipo de dado de valor de decimal para double em todo o back end e criei a função de baixar parcelas no front. Aida falta fazer o bloqueio de pagamentos de parcelas de clientes distintos

+++ MVConsultoria.web/Dtos/ClienteDto.cs
@@ -8,7 +8,7 @@ public class ClienteDto
     public string Telefone { get; set; }
 
     public DateTime DiaDePagamento { get; set; }
-    public decimal LimiteDeCredito { get; set; }
-    public decimal LimiteDisponivel { get; set; }
+    public double LimiteDeCredito { get; set; }
+    public double LimiteDisponivel { get; set; }
     public bool Bloqueado { get; set; }
 }
